pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "variation generator"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "coverage>=7.8.0",
    "flask>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pytest-cov>=6.1.1",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "validators>=0.35.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.5",
    "pytest>=8.3.5",
]